+++ designator-span.py
@@ -50,7 +50,7 @@ from collections import defaultdict
 INVALID = re.compile(r'[\x00-\x1f]|[/\\:*?"<>|,()\'!?、，。；：！？（）【】《》]')
 
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 
 def stamp():